+++ c3smembership/tests/test_models.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8  -*-
 # import os
-from datetime import date
+from datetime import(
+    date,
+    datetime,
+)
 from decimal import Decimal as D
 from decimal import InvalidOperation
 from pyramid import testing
@@ -599,7 +602,7 @@ class Dues15InvoiceModelTests(unittest.TestCase):
             dues1 = Dues15Invoice(
                 invoice_no=1,
                 invoice_no_string=u'C3S-dues15-0001',
-                invoice_date=date.today(),
+                invoice_date=date(2015, 10, 01),
                 invoice_amount=unicode(D('-37.50').to_eng_string()),
                 member_id=1,
                 membership_no=1,
@@ -607,6 +610,70 @@ class Dues15InvoiceModelTests(unittest.TestCase):
                 token=u'ABCDEFGH',
             )
             DBSession.add(dues1)
+
+            dues2 = Dues15Invoice(
+                invoice_no=2,
+                invoice_no_string=u'C3S-dues15-0002-S',
+                invoice_date=date(2015, 10, 02),
+                invoice_amount=unicode(D('17.25').to_eng_string()),
+                member_id=1,
+                membership_no=1,
+                email=u'uat.yes@c3s.cc',
+                token=u'fa4wfjlasjfd',
+            )
+            dues2.is_reversal = True
+            DBSession.add(dues2)
+
+            dues3 = Dues15Invoice(
+                invoice_no=3,
+                invoice_no_string=u'C3S-dues15-0003',
+                invoice_date=date(2015, 11, 25),
+                invoice_amount=unicode(D('74.58').to_eng_string()),
+                member_id=1,
+                membership_no=1,
+                email=u'uat.yes@c3s.cc',
+                token=u'Jleifjsw9e',
+            )
+            DBSession.add(dues3)
+
+            dues4 = Dues15Invoice(
+                invoice_no=4,
+                invoice_no_string=u'C3S-dues15-0004-S',
+                invoice_date=date(2015, 11, 27),
+                invoice_amount=unicode(D('23.85').to_eng_string()),
+                member_id=1,
+                membership_no=1,
+                email=u'uat.yes@c3s.cc',
+                token=u'f348h98sdf',
+            )
+            dues4.is_reversal = True
+            DBSession.add(dues4)
+
+            dues5 = Dues15Invoice(
+                invoice_no=5,
+                invoice_no_string=u'C3S-dues15-0005',
+                invoice_date=date(2015, 11, 29),
+                invoice_amount=unicode(D('12.89').to_eng_string()),
+                member_id=1,
+                membership_no=1,
+                email=u'uat.yes@c3s.cc',
+                token=u'sgdfoiddfg',
+            )
+            DBSession.add(dues5)
+
+            dues6 = Dues15Invoice(
+                invoice_no=6,
+                invoice_no_string=u'C3S-dues15-0006-S',
+                invoice_date=date(2015, 11, 30),
+                invoice_amount=unicode(D('77.79').to_eng_string()),
+                member_id=1,
+                membership_no=1,
+                email=u'uat.yes@c3s.cc',
+                token=u'3o948n',
+            )
+            dues6.is_reversal = True
+            DBSession.add(dues6)
+
             DBSession.flush()
 
     def tearDown(self):
@@ -619,7 +686,7 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         test get_all
         '''
         res = Dues15Invoice.get_all()
-        self.assertEqual(len(res), 1)
+        self.assertEqual(len(res), 6)
 
     def test_get_by_invoice_no(self):
         '''
@@ -640,7 +707,7 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         test get_max_invoice_no
         '''
         res = Dues15Invoice.get_max_invoice_no()
-        self.assertEqual(res, 1)
+        self.assertEqual(res, 6)
 
     def test_check_for_existing_dues15_token(self):
         """
@@ -652,6 +719,21 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         res2 = Dues15Invoice.check_for_existing_dues15_token(
             u'ABCDEFGHIK0000000000')
         self.assertEqual(res2, False)
+
+    def test_get_monthly_stats(self):
+        """
+        Test get_monthly_stats.
+        """
+        stats = Dues15Invoice.get_monthly_stats()
+        self.assertEqual(len(stats), 2)
+        self.assertEqual(stats[0]['month'], datetime(2015, 10, 1))
+        self.assertAlmostEqual(stats[0]['normal_invoices_sum'], D('-37.50'))
+        self.assertAlmostEqual(stats[0]['reversal_invoices_sum'], D('17.25'))
+        self.assertEqual(stats[1]['month'], datetime(2015, 11, 1))
+        self.assertAlmostEqual(stats[1]['normal_invoices_sum'], D('87.47'))
+        self.assertAlmostEqual(stats[1]['reversal_invoices_sum'], D('101.64'))
+
+
 
     def test_decimality(self):
         """
@@ -678,7 +760,7 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         self.session.rollback()
 
         res = Dues15Invoice.get_all()
-        self.assertEqual(len(res), 1)
+        self.assertEqual(len(res), 6)
 
         # try to make another invoice with the same string
         def trigger_IntegrityError2():
@@ -700,13 +782,13 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         self.session.rollback()
 
         res = Dues15Invoice.get_all()
-        self.assertEqual(len(res), 1)
+        self.assertEqual(len(res), 6)
 
         # try to make another invoice with a non-decimal amount
         # InvalidOperation: Invalid literal for Decimal: '-37.50.20'
         def trigger_InvalidOperation():
             dues2 = Dues15Invoice(
-                invoice_no=2,
+                invoice_no=5,
                 invoice_no_string=u'C3S-dues15-0002',
                 invoice_date=date.today(),
                 invoice_amount=unicode(D('-37.50.20').to_eng_string()),
@@ -724,11 +806,11 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         self.session.rollback()
 
         res = Dues15Invoice.get_all()
-        self.assertEqual(len(res), 1)
+        self.assertEqual(len(res), 6)
 
         # now really store a new Dues15Invoice
         dues3 = Dues15Invoice(
-            invoice_no=2,
+            invoice_no=7,
             invoice_no_string=u'C3S-dues15-0002',
             invoice_date=date.today(),
             # invoice_amount=unicode(D('-37.50').to_eng_string()),
@@ -742,6 +824,6 @@ class Dues15InvoiceModelTests(unittest.TestCase):
         DBSession.flush()
 
         res = Dues15Invoice.get_all()
-        self.assertEqual(len(res), 2)
-        self.assertEqual(dues3.id, 2)
+        self.assertEqual(len(res), 7)
+        self.assertEqual(dues3.id, 7)
         # print(type(dues3.invoice_amount))